requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-ckeditor==5.7.1
django-cleanup==4.0.0
django-embed-video==1.2.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
idna==2.8
mysqlclient==1.4.2.post1
Pillow==6.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
